Minor updates to description
I think we should have something about building design culture at Sourcegraph, this might help identify people interested and prepared to work in this kind of environment. 

I think it's better not to specify design tools, to show the applicant that we understand that many tools are popular and that switching a new tool won't be an issue for a professional designer with experience.